+++ docs/form-render/schema/basic.ts
@@ -1,0 +1,23 @@
+export default {
+  type: 'object',
+  column: 3,
+  displayType: 'row',
+  properties: {
+    input1: {
+      title: 'Field A',
+      type: 'string'
+    },
+    input2: {
+      title: 'Field B',
+      type: 'string'
+    },
+    input3: {
+      title: 'Field C',
+      type: 'string'
+    },
+    input4: {
+      title: 'Field D',
+      type: 'string'
+    }
+  }
+};